Move QueryPramDef to internal
